+++ packages/expo-router/src/testing-library/index.tsx
@@ -2,7 +2,7 @@ import './expect';
 import './mocks';
 
 import { NavigationState, PartialState } from '@react-navigation/native';
-import { act, render, RenderResult, screen } from '@testing-library/react-native';
+import { render, RenderResult } from '@testing-library/react-native';
 import React from 'react';
 
 import { MockContextConfig, getMockConfig, getMockContext } from './mock-config';
@@ -10,7 +10,7 @@ import { ExpoRoot } from '../ExpoRoot';
 import { getPathFromState } from '../fork/getPathFromState';
 import { ExpoLinkingOptions } from '../getLinkingConfig';
 import { store } from '../global-state/router-store';
-import { router } from '../imperative-api';
+import { ResultState } from '../exports';
 
 // re-export everything
 export * from '@testing-library/react-native';
@@ -69,62 +69,39 @@ export function renderRouter(
    */
   store.subscribeToRootState(() => jest.runOnlyPendingTimers());
 
+  /**
+   * There maybe additional state updates that occur outside of the initial render cycle.
+   * To avoid the user having to call `act` multiple times, we will just manually update the state here.
+   */
+  const updateRouterState = () => {
+    if (store.navigationRef.isReady()) {
+      const currentState = store.navigationRef.getRootState() as unknown as ResultState;
+      if (store.rootState !== currentState) {
+        store.updateState(currentState);
+      }
+    }
+  };
+
   return Object.assign(result, {
     getPathname(this: RenderResult): string {
+      updateRouterState();
       return store.routeInfoSnapshot().pathname;
     },
     getSegments(this: RenderResult): string[] {
+      updateRouterState();
       return store.routeInfoSnapshot().segments;
     },
     getSearchParams(this: RenderResult): Record<string, string | string[]> {
+      updateRouterState();
       return store.routeInfoSnapshot().params;
     },
     getPathnameWithParams(this: RenderResult): string {
+      updateRouterState();
       return getPathFromState(store.rootState!, store.linking!.config);
     },
     getRouterState(this: RenderResult) {
+      updateRouterState();
       return store.rootStateSnapshot();
     },
   });
 }
-
-export const testRouter = {
-  /** Navigate to the provided pathname and the pathname */
-  navigate(path: string) {
-    act(() => router.navigate(path));
-    expect(screen).toHavePathnameWithParams(path);
-  },
-  /** Push the provided pathname and assert the pathname */
-  push(path: string) {
-    act(() => router.push(path));
-    expect(screen).toHavePathnameWithParams(path);
-  },
-  /** Replace with provided pathname and assert the pathname */
-  replace(path: string) {
-    act(() => router.replace(path));
-    expect(screen).toHavePathnameWithParams(path);
-  },
-  /** Go back in history and asset the new pathname */
-  back(path?: string) {
-    expect(router.canGoBack()).toBe(true);
-    act(() => router.back());
-    if (path) {
-      expect(screen).toHavePathnameWithParams(path);
-    }
-  },
-  /** If there's history that supports invoking the `back` function. */
-  canGoBack() {
-    return router.canGoBack();
-  },
-  /** Update the current route query params and assert the new pathname */
-  setParams(params: Record<string, string>, path?: string) {
-    router.setParams(params);
-    if (path) {
-      expect(screen).toHavePathnameWithParams(path);
-    }
-  },
-  /** If there's history that supports invoking the `back` function. */
-  dismissAll() {
-    act(() => router.dismissAll());
-  },
-};